all post endpoints are good

+++ src/main/java/com/polytech/palmipedewebapp/requests/NidCreationRequest.java
@@ -2,31 +2,32 @@ package com.polytech.palmipedewebapp.requests;
 
 public class NidCreationRequest {
 
-    private Float idBatiment;
-    private Float idAntenne;
-    private Float idBalance;
+    private Long idBatiment;
+    private Long idAntenne;
+    private Long idBalance;
 
-    public Float getIdBatiment() {
+    public Long getIdBatiment() {
         return idBatiment;
     }
 
-    public void setIdBatiment(Float idBatiment) {
+    public void setIdBatiment(Long idBatiment) {
         this.idBatiment = idBatiment;
     }
 
-    public Float getIdAntenne() {
+    public Long getIdAntenne() {
+
         return idAntenne;
     }
 
-    public void setIdAntenne(Float idAntenne) {
+    public void setIdAntenne(Long idAntenne) {
         this.idAntenne = idAntenne;
     }
 
-    public Float getIdBalance() {
+    public Long getIdBalance() {
         return idBalance;
     }
 
-    public void setIdBalance(Float idBalance) {
+    public void setIdBalance(Long idBalance) {
         this.idBalance = idBalance;
     }
 }
